Multiple Changes and implementation of first strategies
changed portfolio money to double
added comission fee parameter
implemented SMA strategy
implemented Bollinger Band Strategy
added Observer for Strategys
added Statistics Object
added text output in /Output/ as csv for further evaluation ie with
matlab

+++ EC910GroupProject/src/org/com/essex/ec910/artificialstockmarket/trader/Portfolio.java
@@ -9,7 +9,7 @@ public class Portfolio {
 
 
 	private int shares;   // number of shares in trader's portfolio 	
-	private int money;    // trader's capital
+	private double money;    // trader's capital
 		
 	/**
 	 * Constructor
@@ -29,11 +29,11 @@ public class Portfolio {
 		this.shares = shares;
 	}
 	
-	public int getMoney() {
+	public double getMoney() {
 		return money;
 	}
 	
-	public void setMoney(int money) {
+	public void setMoney(double money) {
 		this.money = money;
 	}
